+++ lib/clean_results.dart
@@ -1,0 +1,11 @@
+class CleanResults {
+  final int projectsFound;
+  final int projectsCleaned;
+  final int spaceFreed;
+
+  CleanResults({
+    required this.projectsFound,
+    required this.projectsCleaned,
+    required this.spaceFreed,
+  });
+}