+++ lib/flutter_project.dart
@@ -1,0 +1,9 @@
+class FlutterProject {
+  final String path;
+  final int buildSize;
+
+  FlutterProject({
+    required this.path,
+    required this.buildSize,
+  });
+}